wkbotfeeder.py: add configuration parameter 'target_step' to determine when a bot build is successful
Bots can now configure a parameter 'target_step' to watch a specific
build step and determine if they're successful.

For instance, a bot that builds and run tests, like the JSC bots, can
watch only the 'compile-webkit' step and determine whether a build was
successful if that step returned 'compiled'. Example:

"builders": {
    "JSCOnly Linux MIPS32el Release",
    "last_buildjob_url_squema": "https://build.webkit.org/builders/%(builder_name)s/builds/%(last_buildjob)s",
    "builds_url_squema": "https://build.webkit.org/json/builders/%(builder_name)s/builds?select=-2&as_text=1",
    "builder_name": 'JSCOnly Linux MIPS32el Release',
    "only_failures": False
    "target_step": {
        "name": "compile-webkit",
        "text": "compiled"
    }
}